requirements.txt
Django==1.10.2
djangorestframework==3.4.7
Pillow==3.4.1
psycopg2==2.6.2
